+++ src/type/savegame.ts
@@ -1,0 +1,7 @@
+import Version from "../class/SaveGameReader";
+
+interface SaveGameMod {
+    name: string,
+    version: Version,
+    checksum: number,
+}