+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.0 - 2020-06-12
+
+### Added
+
+- Link to the Discord server to the footer.
+
+### Changed
+
+- Adding a setting now requires a savegame to be selected, not the mod-list.json file anymore.
+- Arrangement of boxes and explanations on settings/new page.
+
 ## 1.0.0 - 2020-05-22
 
 - Initial release of the portal frontend project.